+++ circuit-shared-elements/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=circuit-sharedelements
+POM_NAME=Circuit (Shared Elements)
+POM_DESCRIPTION=Circuit Shared Elements
