+++ samples/star/src/androidUnitTest/kotlin/com/slack/circuit/star/petlist/PetListUiTest.kt
@@ -16,8 +16,8 @@ import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
 import coil.annotation.ExperimentalCoilApi
 import coil3.test.FakeImage
-import com.slack.circuit.foundation.PreviewSharedElementTransitionLayout
 import com.slack.circuit.sample.coil.test.CoilRule
+import com.slack.circuit.sharedelements.PreviewSharedElementTransitionLayout
 import com.slack.circuit.star.common.Strings
 import com.slack.circuit.star.db.Gender.MALE
 import com.slack.circuit.star.db.Size.SMALL
